services/translation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.26.0
cachetools==5.3.2

